refactor: remove old `features/Chat/Thread`
