--- refact-agent/gui/src/features/Chat/Thread/index.ts
@@ -1,1 +1,0 @@
-export * from "./types";
