+++ refact-agent/gui/src/services/graphql/queriesAndMutationsApi.ts
@@ -40,7 +40,7 @@ import {
 import { type RootState } from "../../app/store";
 import { setThreadFtId } from "../../features/ThreadMessages";
 import { Tool } from "../refact/tools";
-import { IntegrationMeta } from "../../features/Chat";
+import type { IntegrationMeta } from "../../features/ThreadMessages";
 
 // async function fetchAppSearchableId(apiKey: string, port: number) {
 //   const appIdUrl = `http://127.0.0.1:${port}/v1/get-app-searchable-id`;
