+++ refact-agent/gui/src/services/refact/links.ts
@@ -1,9 +1,18 @@
 import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react";
 import { RootState } from "../../app/store";
-import { ChatMessage, ChatMessages } from "./types";
-import { formatMessagesForLsp } from "../../features/Chat/Thread/utils";
+import {
+  ChatMessage,
+  ChatMessages,
+  isAssistantMessage,
+  isDiffMessage,
+  isToolMessage,
+  isUserMessage,
+  LspChatMode,
+} from "./types";
 import { CHAT_COMMIT_LINK_URL, CHAT_LINKS_URL } from "./consts";
-import { LspChatMode } from "../../features/Chat";
+
+import { LspChatMessage, LSPUserMessage } from "./chat";
+
 // useful for forcing specific links
 // import { STUB_LINKS_FOR_CHAT_RESPONSE } from "../../__fixtures__";
 
@@ -226,4 +235,56 @@ function isCommitResponse(json: unknown): json is CommitResponse {
   if (!Array.isArray(json.error_log)) return false;
   // TODO: type check the arrays if we use the data anywhere.
   return true;
+}
+
+export function formatMessagesForLsp(messages: ChatMessages): LspChatMessage[] {
+  return messages.reduce<LspChatMessage[]>((acc, message) => {
+    if (isUserMessage(message)) {
+      const { ftm_role, ftm_content, ...rest } = message;
+      const msg: LSPUserMessage = {
+        ...rest,
+        role: ftm_role,
+        content: ftm_content,
+      };
+      return acc.concat([msg]);
+    }
+
+    if (isAssistantMessage(message)) {
+      return acc.concat([
+        {
+          role: message.ftm_role,
+          content: message.ftm_content,
+          tool_calls: message.ftm_tool_calls ?? undefined,
+          thinking_blocks: message.thinking_blocks ?? undefined,
+          finish_reason: message.finish_reason,
+          usage: message.usage,
+        },
+      ]);
+    }
+
+    if (isToolMessage(message)) {
+      return acc.concat([
+        {
+          role: "tool",
+          content: message.ftm_content,
+          tool_call_id: message.ftm_call_id,
+        },
+      ]);
+    }
+
+    if (isDiffMessage(message)) {
+      const diff = {
+        role: message.ftm_role,
+        content: JSON.stringify(message.ftm_content),
+        tool_call_id: message.tool_call_id,
+      };
+      return acc.concat([diff]);
+    }
+
+    const ftm_content =
+      typeof message.ftm_content === "string"
+        ? message.ftm_content
+        : JSON.stringify(message.ftm_content);
+    return [...acc, { role: message.ftm_role, content: ftm_content }];
+  }, []);
 }
